Update movie and series detail screens with improved UI and metadata formatting
- Redesign `MovieScreen` and `DetailsScreen` with a focus on structured metadata (rating, genre, release date, runtime, status).
- Implement specialized formatting for runtimes, release dates, and series seasons/episodes counts.
- Add `Image.toRequest()` extension using Coil3 for standardized image loading with placeholders and crossfade.
- Introduce `BaseScreens.kt` with reusable `ErrorScreen` and `LoadingScreen` components.
- Update `CastWidget` with improved styling and support for episode counts using plurals.
- Refactor domain models (`BaseDetails`, `MovieDetail`, `SeriesDetail`) to use more appropriate data types (e.g., `Int` for IDs/counts, `Float` for popularity/ratings).
- Enhance `ExpandableText` with localized resources and adjustable minimum lines.
- Add `FormatConstant` and `SymbolConstant` for consistent string manipulation.
- Update TMDB DTOs to reflect nullable fields and domain model changes.

+++ core/domain/src/main/java/com/vroff/domain/model/tmdb/search/SearchResult.kt
@@ -22,7 +22,7 @@ data class SearchResult(
     val genres: List<Genre>,
     val popularity: Double,
     val firstAirDate: String = "",
-    val voteAverage: Double = 0.0,
+    val voteAverage: Float = 0.0f,
     val voteCount: Long = 0,
     val originCountry: List<String>,
     val gender: Gender = Gender.NOT_SET,
